PROTOCOL: Flush any pending line numbers when idle
There appears to be a race condition that prevents the reporting system
from actually reporting all completed lines.  I have a sneaking
suspicion that it was introduced while trying to combat lost linenumbers
on 0-length moves / alarms, and has existed for quite some time, but was
being masked by the streaming status reports.  When we moved back to
reactive/line end reports, this condition cropped back up as we weren't
constantly pinging the line number report.