PROTOCOL: Remove program manager (#105)
This removes the 'program manager' functionality that was added a while
back in an effort to buffer larger chunks of gcode.  Unfortunately, this
has put us in a bit of a bind and caused a lot of headaches related to
gcode failing to execute and has introduced a number of subtle race
conditions with our new cutting system.  By switching back, it makes
everything a bit more synchronous as we only get positive
confirmation ('ok') when a line has been added to the line buffer, NOT
when it has been added to the internal data buffer.