SYSTEM: Replaces '$P' with potentiometer commands
We no longer use the old force sensors, so this made sense to just take
over and replace with the new digital potentiometer controls.
Currently, the $P0S and $P0W commands are supported (Store / Write
respectively).  We can also support a 'read' command, but the effect is
sort of implicit as the voltage change at the load cell in pin can be
measured to see the effect (which is arguably more important)