DEVICES: Refactor SRAM stuff (#119)
This is a follow-up to #118. Not tested (I don't know how to run this code), though it does compile.

I passed in the length of the array to _sram_transact_helper, as discussed in #118 (comment)
I refactored sram_read_mode and sram_set_mode as discussed in #118 (comment), and passed in the length of the array in a similar manner.